Moved storage writer and clean up #102 #351
